requirements.txt
# GUI Framework
PySide6>=6.6.0

# Excel & PDF
openpyxl>=3.1.2
reportlab>=4.0.7

# AI/LLM Integration
anthropic>=0.18.0
openai>=1.12.0
python-dotenv>=1.0.0

# Document Processing
pdfplumber>=0.10.0
pytesseract>=0.3.10
Pillow>=10.0.0
python-magic-bin>=0.4.14

# Utilities
requests>=2.31.0
